DL-12453: Play 3.0 upgrade and adressing integrationTest scope
