+++ conf/prod.routes
@@ -1,4 +1,3 @@
 # Add all the application routes to the app.routes file
 ->         /ated                      app.Routes
 ->         /                          health.Routes
-GET        /admin/metrics             @com.kenshoo.play.metrics.MetricsController.metrics
