+++ it/test/connectors/AddressLookupConnectorISpec.scala
@@ -1,10 +1,27 @@
-package connectors
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import helpers.IntegrationBase
-import helpers.stubs.{AddressLookupStub, AuthAudit}
+package test.connectors
+
+import connectors.AddressLookupConnector
 import models.{AddressLookup, AddressLookupCountry, AddressLookupRecord, AddressSearchResult}
 import play.api.test.Injecting
 import play.api.test.Helpers._
+import test.helpers.IntegrationBase
+import test.helpers.stubs.{AddressLookupStub, AuthAudit}
 import uk.gov.hmrc.http.HeaderCarrier
 
 class AddressLookupConnectorISpec extends IntegrationBase with AuthAudit with Injecting {
