+++ it/test/helpers/IntegrationBase.scala
@@ -1,11 +1,23 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-package helpers
+package test.helpers
 
 import com.github.tomakehurst.wiremock.client.WireMock._
 import com.github.tomakehurst.wiremock.stubbing.StubMapping
-import helpers.application.IntegrationApplication
-import helpers.stubs.GGLoginStub
-import helpers.wiremock.WireMockSetup
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.ws.WSRequest
@@ -13,6 +25,9 @@ import play.api.test.{DefaultAwaitTimeout, FutureAwaits}
 import uk.gov.hmrc.http.HeaderNames
 import play.api.libs.ws.WSClient
 import play.api.http.{HeaderNames => HN}
+import test.helpers.application.IntegrationApplication
+import test.helpers.stubs.GGLoginStub
+import test.helpers.wiremock.WireMockSetup
 
 trait IntegrationBase extends PlaySpec
   with BeforeAndAfterEach
