+++ project/AppDependencies.scala
@@ -6,37 +6,29 @@ object AppDependencies {
   import play.core.PlayVersion
   import play.sbt.PlayImport._
 
+  val jacksonVersion = "2.17.0"
+  val bootstrapVersion = "8.5.0"
+
   val compile: Seq[ModuleID] = Seq(
     ws,
-    "uk.gov.hmrc"       %% "bootstrap-frontend-play-28"  % "7.23.0",
-    "uk.gov.hmrc"       %% "play-partials"               % "8.4.0-play-28",
-    "uk.gov.hmrc"       %% "domain"                      % "8.3.0-play-28",
-    "uk.gov.hmrc"       %% "http-caching-client"         % "10.0.0-play-28",
-    "uk.gov.hmrc"       %% "play-frontend-hmrc-play-28"  % "8.5.0"
+    "uk.gov.hmrc"       %% "bootstrap-frontend-play-30"  % bootstrapVersion,
+    "uk.gov.hmrc"       %% "play-partials-play-30"       % "9.1.0",
+    "uk.gov.hmrc"       %% "domain-play-30"              % "9.0.0",
+    "uk.gov.hmrc"       %% "http-caching-client-play-30" % "11.2.0",
+    "uk.gov.hmrc"       %% "play-frontend-hmrc-play-30"  % "8.5.0"
   )
 
-  trait TestDependencies {
-    lazy val scope: String = "it,test"
-    lazy val test: Seq[ModuleID] = ???
-  }
-
-  object Test {
-    def apply(): Seq[ModuleID] = new TestDependencies {
-      override lazy val test: Seq[ModuleID] = Seq(
-        "uk.gov.hmrc"                  %% "bootstrap-test-play-28"      % "7.23.0"            % scope,
-        "org.scalatestplus.play"       %% "scalatestplus-play"          % "5.1.0"             % scope,
-        "org.jsoup"                    %  "jsoup"                       % "1.17.2"            % scope,
-        "com.typesafe.play"            %% "play-test"                   % PlayVersion.current % scope,
-        "org.scalatestplus"            %% "mockito-4-11" % "3.2.17.0"   % scope,
-        "org.mockito" % "mockito-core" % "5.5.0"                        % scope,
-        "com.github.tomakehurst"       %  "wiremock-jre8"               % "2.35.1"            % scope,
-        "com.fasterxml.jackson.module" %% "jackson-module-scala"        % "2.15.2"            % scope
+ val test: Seq[ModuleID] = Seq(
+        "uk.gov.hmrc"                  %% "bootstrap-test-play-30"      % bootstrapVersion    % "test",
+        "org.jsoup"                    %  "jsoup"                       % "1.17.2"            % "test",
+        "org.scalatestplus"            %%  "mockito-4-11"               % "3.2.18.0"          % "test",
+        "org.mockito"                  %  "mockito-core"                % "5.11.0"            % "test",
+        "com.fasterxml.jackson.module" %% "jackson-module-scala"        % jacksonVersion      % "test"
       )
-    }.test
-  }
 
-  val jacksonVersion = "2.16.1"
-  val jacksonDatabindVersion = "2.16.1"
+  val itDependencies: Seq[ModuleID] = Seq(
+        "org.wiremock"                 %  "wiremock"                    % "3.3.1"             % Test
+  )
 
   val jacksonOverrides = Seq(
     "com.fasterxml.jackson.core" % "jackson-core",
@@ -46,7 +38,7 @@ object AppDependencies {
   ).map(_ % jacksonVersion)
 
   val jacksonDatabindOverrides = Seq(
-    "com.fasterxml.jackson.core" % "jackson-databind" % jacksonDatabindVersion
+    "com.fasterxml.jackson.core" % "jackson-databind" % jacksonVersion
   )
 
   val akkaSerializationJacksonOverrides = Seq(
@@ -55,5 +47,5 @@ object AppDependencies {
     "com.fasterxml.jackson.module" %% "jackson-module-scala",
   ).map(_ % jacksonVersion)
 
-  def apply(): Seq[ModuleID] = compile ++ jacksonDatabindOverrides ++ jacksonOverrides ++ akkaSerializationJacksonOverrides ++ Test()
+  def apply(): Seq[ModuleID] = compile ++ jacksonDatabindOverrides ++ jacksonOverrides ++ akkaSerializationJacksonOverrides ++ test
 }
